requirements.txt
moviepy
imagebind @ git+https://github.com/algotech-ai/ImageBind.git@d84fd6e2e50d44242b6ee1a854a050393453b86f
ffmpeg-python